feat!: add toast_body and rm warning, error, confirmation toast
