+++ example/lib/pages/toast_page.dart
@@ -26,7 +26,7 @@ class ToastPage extends StatelessWidget {
                   child: SBBTertiaryButtonLarge(
                     label: 'Show Toast - Short (Default)',
                     onPressed: () {
-                      sbbToast.warning(
+                      sbbToast.show(
                         message: 'const SBBToast.durationShort: ${SBBToast.durationShort.inSeconds} seconds',
                       );
                     },
@@ -51,7 +51,7 @@ class ToastPage extends StatelessWidget {
                   child: SBBTertiaryButtonLarge(
                     label: 'Show Toast - 5 Seconds',
                     onPressed: () {
-                      sbbToast.error(
+                      sbbToast.show(
                         message: 'This is a Toast with a duration of 5 seconds.',
                         duration: const Duration(seconds: 5),
                       );
